requirements.txt
fastapi
uvicorn
python-dotenv
numpy==2.3.5
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.30,<0.4.0
langchain-google-genai
pinecone
tiktoken
twilio
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
pydantic
python-multipart
jinja2
